+++ modules/db/db.module.spec.js
@@ -9,25 +9,38 @@ const
 module.exports = {
 	testMe: function() {
 		const instance = this
+		this.config.postgreSQL.logging = false
 		describe('core.modules.db', function() {
+			it('should execute testConnectToDB successfully', function() {
+				instance.testConnectToDB()
+				assert(true)
+			})
 			it('should execute testLoadComponents successfully', function() {
 				instance.testLoadComponents()
 				assert(true)
 			})
+			it('should execute testCreateAssociations successfully', function() {
+				instance.testCreateAssociations()
+				assert(true)
+			})
+			it('should execute testSetComponentsProperties successfully', function() {
+				instance.testSetComponentsProperties()
+				assert(true)
+			})
 		})
 	},
-	testLoadComponents: function() {
+	testConnectToDB: function() {
 		const instance = this,
 			{config, moduleConfig} = this,
 			originalConfig = JSON.parse(JSON.stringify(config))
 		let changeableInstance = this
-		describe('core.modules.db.loadComponents', function() {
+		describe('core.modules.db.connectToDB', function() {
 			it('should throw an error with the correct message if an invalid dbType is provided', function() {
 				return co(function*() {
 					let didThrowAnError = false
 					changeableInstance.config.db.dbType = 'invalidDBType'
 					try {
-						yield instance.loadComponents()
+						yield instance.connectToDB()
 					} catch(e) {
 						didThrowAnError = e && (e.customMessage === 'Invalid dbType.')
 					}
@@ -44,7 +57,7 @@ module.exports = {
 						let didThrowAnError = false
 						postgreSQLConfig.host = `absolutelyFakeHostAddress_${moment.utc().valueOf()}`
 						try {
-							yield instance.loadComponents()
+							yield instance.connectToDB()
 						} catch(e) {
 							didThrowAnError = e && (e.name === 'SequelizeHostNotFoundError')
 						}
@@ -58,7 +71,7 @@ module.exports = {
 						let didThrowAnError = false
 						postgreSQLConfig.port = `absolutelyFakeServerPort_${moment.utc().valueOf()}`
 						try {
-							yield instance.loadComponents()
+							yield instance.connectToDB()
 						} catch(e) {
 							didThrowAnError = e && (e.name === 'RangeError')
 						}
@@ -72,7 +85,7 @@ module.exports = {
 						let didThrowAnError = false
 						postgreSQLConfig.user = `absolutelyFakeUsername_${moment.utc().valueOf()}`
 						try {
-							yield instance.loadComponents()
+							yield instance.connectToDB()
 						} catch(e) {
 							didThrowAnError = e && (e.name === 'SequelizeConnectionError')
 						}
@@ -86,7 +99,7 @@ module.exports = {
 						let didThrowAnError = false
 						postgreSQLConfig.password = `absolutelyFakePassword_${moment.utc().valueOf()}`
 						try {
-							yield instance.loadComponents()
+							yield instance.connectToDB()
 						} catch(e) {
 							didThrowAnError = e && (e.name === 'SequelizeConnectionError')
 						}
@@ -101,7 +114,7 @@ module.exports = {
 						let didThrowAnError = false
 						postgreSQLConfig.database = `absolutelyFakeDatabaseName_${moment.utc().valueOf()}`
 						try {
-							yield instance.loadComponents()
+							yield instance.connectToDB()
 						} catch(e) {
 							didThrowAnError = e && (e.name === 'SequelizeConnectionError')
 						}
@@ -112,11 +125,219 @@ module.exports = {
 				})
 				it('should execute successfully if all paramters are correct', function() {
 					return co(function*() {
-						yield instance.loadComponents()
+						yield instance.connectToDB()
 						assert(true)
 						return true
 					})
 				})
+			})
+		})
+	},
+	testLoadComponents: function() {
+		const instance = this,
+			{config, moduleConfig} = this,
+			originalConfig = JSON.parse(JSON.stringify(config))
+		let changeableInstance = this
+		describe('core.modules.db.loadComponents', function() {
+			it('should throw an error if moduleConfig.modulePath is undefined', function() {
+				return co(function*() {
+					let didThrowAnError = false
+					changeableInstance.config.db.modulePath = undefined
+					try {
+						yield instance.loadComponents()
+					} catch(e) {
+						didThrowAnError = true
+					}
+					changeableInstance.config.db.modulePath = originalConfig.db.modulePath
+					assert(didThrowAnError)
+					return true
+				})
+			})
+			it('should throw an error if moduleConfig.modulePath is null', function() {
+				return co(function*() {
+					let didThrowAnError = false
+					changeableInstance.config.db.modulePath = null
+					try {
+						yield instance.loadComponents()
+					} catch(e) {
+						didThrowAnError = true
+					}
+					changeableInstance.config.db.modulePath = originalConfig.db.modulePath
+					assert(didThrowAnError)
+					return true
+				})
+			})
+			it('should throw an error if moduleConfig.modulePath does not point to a valid directory', function() {
+				return co(function*() {
+					let didThrowAnError = false
+					changeableInstance.config.db.modulePath = 'notAValidDirectoryByFar'
+					try {
+						yield instance.loadComponents()
+					} catch(e) {
+						didThrowAnError = true
+					}
+					changeableInstance.config.db.modulePath = originalConfig.db.modulePath
+					assert(didThrowAnError)
+					return true
+				})
+			})
+			it('should throw an error with the correct message if a component\'s class does not contain a model object', function() {
+				return co(function*() {
+					let didThrowAnError = false,
+						componentName = `absolutelyFakeModule_${moment.utc().valueOf()}`,
+						componentPath = path.join(moduleConfig.modulePath, componentName)
+					yield fs.mkdirp(componentPath)
+					let fd = yield fs.open(path.join(componentPath, 'index.js'), 'w')
+					yield fs.writeFile(fd, 'class DBComponent {constructor() {}}; module.exports = DBComponent;')
+					yield fs.close(fd)
+					try {
+						yield instance.loadComponents()
+					} catch(e) {
+						didThrowAnError = e && (e.customMessage === `DB module component "${componentName}" loaded, does not have a valid model.`)
+					}
+					yield fs.remove(componentPath)
+					assert(didThrowAnError)
+					return true
+				})
+			})
+			it('should throw an error with the correct message if a component has a spec file, but it is invalid', function() {
+				return co(function*() {
+					let didThrowAnError = false,
+						componentName = `absolutelyFakeModule_${moment.utc().valueOf()}`,
+						componentPath = path.join(moduleConfig.modulePath, componentName)
+					yield fs.mkdirp(componentPath)
+					let fd = yield fs.open(path.join(componentPath, 'index.js'), 'w')
+					yield fs.writeFile(fd, 'class DBComponent {constructor() {this.model = function() {}}}; module.exports = DBComponent;')
+					yield fs.close(fd)
+					fd = yield fs.open(path.join(componentPath, 'index.spec.js'), 'w')
+					yield fs.writeFile(fd, 'invalid')
+					yield fs.close(fd)
+					try {
+						yield instance.loadComponents()
+					} catch(e) {
+						didThrowAnError = e && (e.customMessage === `Invalid spec file for DB module component "${componentName}".`)
+					}
+					yield fs.remove(componentPath)
+					assert(didThrowAnError)
+					return true
+				})
+			})
+			it('should execute successfully if all paramters are correct and doSync isn\'t enabled', function() {
+				return co(function*() {
+					yield instance.loadComponents()
+					assert(true)
+					return true
+				})
+			})
+			it('should have loaded all components after it has executed successfully', function() {
+				return co(function*() {
+					const components = instance.components
+					let moduleDirData = yield fs.readdir(moduleConfig.modulePath),
+						allLoaded = true
+					for (const i in moduleDirData) {
+						let componentName = moduleDirData[i]
+						if ((componentName.indexOf('.') === -1) && !components[componentName]){
+							allLoaded = false
+							break
+						}
+					}
+					assert(allLoaded)
+					return true
+				})
+			})
+			it('should have set the correct componentName for all components after it has executed successfully', function() {
+				const components = instance.components
+				let allHaveCorrectComponentNames = true
+				for (const componentName in components) {
+					const component = components[componentName]
+					if (!component.componentName || (component.componentName !== componentName)) {
+						allHaveCorrectComponentNames = false
+						break
+					}
+				}
+				assert(allHaveCorrectComponentNames)
+			})
+			it('should have set the db property, with the component removed individually to avoid circularization, for all components after it has executed successfully', function() {
+				const components = instance.components
+				let allHaveTheValidDBProperty = true
+				for (const componentName in components) {
+					const component = components[componentName]
+					if (!component.db || component.db.components[componentName]) {
+						allHaveTheValidDBProperty = false
+						break
+					}
+				}
+				assert(allHaveTheValidDBProperty)
+			})
+			it('should have loaded the fieldCaseMap successfully, if a valid one was present in the module directory', function() {
+				return co(function*() {
+					let moduleDirData = yield fs.readdir(moduleConfig.modulePath),
+						hasFieldCaseMap = true
+					for (const i in moduleDirData) {
+						let componentName = moduleDirData[i]
+						if (componentName === 'fieldCaseMap.js') {
+							if (!instance.fieldCaseeMap) {
+								hasFieldCaseMap = false
+							}
+							break
+						}
+					}
+					assert(hasFieldCaseMap)
+					return true
+				})
+			})
+		})
+	},
+	testCreateAssociations: function() {
+		const instance = this,
+			{config, moduleConfig} = this,
+			originalConfig = JSON.parse(JSON.stringify(config))
+		let changeableInstance = this
+		describe('core.modules.db.createAssociations', function() {
+			it.skip('should execute successfully if all paramters are correct and doSync isn\'t enabled', function() {
+				return co(function*() {
+					yield instance.createAssociations()
+					assert(true)
+					return true
+				})
+			})
+			it('should execute successfully if all paramters are correct and doSync is enabled', function() {
+				return co(function*() {
+					// yield instance.migrations.removeAllTables()
+					yield instance.createAssociations(true)
+					assert(true)
+					return true
+				})
+			})
+		})
+	},
+	testSetComponentsProperties: function() {
+		const instance = this,
+			{config, moduleConfig} = this,
+			originalConfig = JSON.parse(JSON.stringify(config))
+		let changeableInstance = this
+		describe('core.modules.db.setComponentsProperties', function() {
+			it('should throw an error with the correct message if the properties argument is not a valid object', function() {
+				let didThrowAnError = false
+				try {
+					instance.setComponentsProperties()
+				} catch(e) {
+					didThrowAnError = e && (e.customMessage === 'Invalid properties object provided.')
+				}
+				assert(didThrowAnError)
+			})
+			it('should set the provided properties to all components correctly', function() {
+				const components = instance.components
+				let allHaveTheCorrectProperties = true
+				instance.setComponentsProperties({test1: 'correct', test2: 'properties'})
+				for (const componentName in components) {
+					const component = components[componentName]
+					if ((component.test1 !== 'correct') || (component.test2 !== 'properties')) {
+						allHaveTheCorrectProperties = false
+						break
+					}
+				}
+				assert(allHaveTheCorrectProperties)
 			})
 		})
 	}