+++ test/config/common.js
@@ -39,9 +39,7 @@ module.exports = {
 	db: {
 		modulePath: path.join(__dirname, '../modules/db'),
 		dbType: 'postgreSQL',
-		seedingOrder: [
-			'modules', 'roles', 'keyAccessPoints', 'users'
-		]
+		schema: 'public'
 	},
 	clients: {
 		site: {