+++ test/modules/db/moduleCategories/index.js
@@ -1,0 +1,43 @@
+'use strict'
+
+const
+	Base = require('../../../../index.js').BaseDBComponent
+
+class Component extends Base {
+	constructor(sequelize, Sequelize, settings) {
+		super(settings)
+
+		this.model = sequelize.define('moduleCategory', {
+			name: {type: Sequelize.STRING, allowNull: false, validate: {notEmpty: true}},
+			description: {type: Sequelize.STRING, allowNull: false, validate: {notEmpty: true}},
+			order: {type: Sequelize.INTEGER, allowNull: false, validate: {min: 0}},
+			icon: {type: Sequelize.STRING, allowNull: false, validate: {notEmpty: true}},
+		}, {
+			indexes: [
+				{unique: true, fields: ['name'], where: {deletedAt: null}},
+				{unique: true, fields: ['order'], where: {deletedAt: null}}
+			],
+			setterMethods: {
+				id: function (value) {
+				}
+			},
+			paranoid: true
+		})
+
+		this.associationsConfig = {
+			modules: {type: 'hasMany', foreignKey: 'categoryId'}
+		}
+
+		this.searchFields = [
+			{field: 'id'},
+			{field: 'name'},
+			{field: 'description'},
+			{field: 'order'},
+			{field: 'createdAt'},
+			{field: 'updatedAt'},
+			{field: 'deletedAt'}
+		]
+	}
+}
+
+module.exports = Component